backend/requirements.txt
fastapi[standard]
uvicorn
openai
pymupdf
python-docx
python-multipart
pydantic
python-dotenv